+++ flattentool/tests/test_docs.py
@@ -90,9 +90,9 @@ def test_examples_in_docs():
                 tests_passed += 1
     # Check that the number of tests were run that we expected
     if sys.version_info[:2] < (3,4):
-        assert tests_passed == 41
-    else:
         assert tests_passed == 42
+    else:
+        assert tests_passed == 43
 
 def _simplify_warnings(lines):
     return '\n'.join([_simplify_line(line) for line in lines.split('\n')])